Refactor login fields from NavigationBar component into LoginForm component hide login fields, login button and signup button when user logs in
