+++ src/components/NavigationBar.jsx
@@ -5,40 +5,16 @@ import {
   Toolbar,
   Typography,
   CssBaseline,
-  Button,
-  TextField
+  Button
 } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import Authentication from "../modules/Authentication";
-import { useSelector, useDispatch } from "react-redux";
+import { useSelector } from "react-redux";
+import LoginForm from "./LoginForm";
 
 const NavigationBar = () => {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
   const { currentUser } = useSelector((state) => state);
   const [showLogin, setShowLogin] = useState(false);
-  const [loginForm, setLoginForm] = useState({});
-
-  const handleChange = (event) => {
-    setLoginForm({
-      ...loginForm,
-      [event.target.name]: event.target.value
-    });
-  };
-
-  const handleEnter = async (event) => {
-    if (event.keyCode === 13) {
-      const response = await Authentication.signIn(loginForm);
-      if (response.success) {
-        dispatch({ type: "SET_CURRENT_USER", payload: response.data });
-      } else {
-        dispatch({
-          type: "SET_ERROR_MESSAGE",
-          payload: response
-        });
-      }
-    }
-  };
 
   return (
     <Box sx={{ flexGrow: 1 }}>
@@ -61,46 +37,30 @@ const NavigationBar = () => {
           >
             Recipe Hub
           </Typography>
-          {showLogin ? (
-            <React.Fragment>
-              <TextField
-                data-cy="email-input"
-                label="Email"
-                name="email"
-                size="small"
-                onChange={handleChange}
-                onKeyDown={handleEnter}
-                variant="outlined"
-                style={{ backgroundColor: "white", borderRadius: "5px" }}
-              />
-              <TextField
-                data-cy="password-input"
-                label="Password"
-                name="password"
-                type="password"
-                size="small"
-                onChange={handleChange}
-                onKeyDown={handleEnter}
-                variant="outlined"
-                style={{ backgroundColor: "white", borderRadius: "5px" }}
-              />
-            </React.Fragment>
+          {currentUser ? (
+            <div data-cy="user-name">{currentUser.name}</div>
           ) : (
             <React.Fragment>
-              <Button
-                color="inherit"
-                data-cy="login-btn"
-                onClick={() => setShowLogin(!showLogin)}
-              >
-                Log In
-              </Button>
-              <Button
-                color="inherit"
-                data-cy="sign-up-btn"
-                onClick={() => navigate("/signup")}
-              >
-                Sign Up
-              </Button>
+              {showLogin ? (
+                <LoginForm />
+              ) : (
+                <React.Fragment>
+                  <Button
+                    color="inherit"
+                    data-cy="login-btn"
+                    onClick={() => setShowLogin(!showLogin)}
+                  >
+                    Log In
+                  </Button>
+                  <Button
+                    color="inherit"
+                    data-cy="sign-up-btn"
+                    onClick={() => navigate("/signup")}
+                  >
+                    Sign Up
+                  </Button>
+                </React.Fragment>
+              )}
             </React.Fragment>
           )}
         </Toolbar>
